Remove old markdown component
