+++ src/components/Documentation/Markdown/index.js
@@ -1,6 +1,6 @@
-import React from 'react'
+import React, { useCallback, useEffect, useRef } from 'react'
+import rehypeReact from 'rehype-react'
 import PropTypes from 'prop-types'
-import ReactMarkdown from 'react-markdown'
 import Collapsible from 'react-collapsible'
 
 import 'github-markdown-css/github-markdown.css'
@@ -20,125 +20,126 @@ import {
   TutorialsWrapper
 } from './styles'
 
-const HtmlRenderer = props => {
-  if (props.tag !== 'details' && props.tag !== 'abbr') {
-    return React.createElement(props.tag, {}, props.children)
-  } else if (props.tag === 'details') {
-    const text = props.children[0].props.children[0]
-    return (
-      <Collapsible trigger={text} transitionTime={200}>
-        {props.children.slice(1)}
-      </Collapsible>
-    )
-  } else if (props.tag === 'abbr') {
-    const text = props.children[0]
-    const key = props.children[0].key
-    return <Tooltip id={key} text={text} />
+const isInsideCodeBlock = elem => {
+  for (let el = elem; el && el !== document; el = el.parentNode) {
+    if (el.tagName === 'PRE') return true
+    if (el.tagName === 'ARTICLE') return false
   }
+  return false
 }
 
-export default class Markdown extends React.PureComponent {
-  constructor() {
-    super()
-    this.touchstartX = 0
-    this.touchendX = 0
-    this.isCodeBlock = false
-  }
+function Details({ children }) {
+  const filteredChildren = children.filter(child => child !== '\n')
 
-  componentDidMount() {
-    document.addEventListener('touchstart', this.onTouchStart, false)
-    document.addEventListener('touchend', this.onTouchEnd, false)
-  }
+  const text = filteredChildren[0].props.children[0]
+  return (
+    <Collapsible trigger={text} transitionTime={200}>
+      {filteredChildren.slice(1)}
+    </Collapsible>
+  )
+}
 
-  componentWillUnmount() {
-    document.removeEventListener('touchstart', this.onTouchStart)
-    document.removeEventListener('touchend', this.onTouchEnd)
-  }
+Details.propTypes = {
+  children: PropTypes.node
+}
 
-  isInsideCodeBlock = elem => {
-    for (let el = elem; el && el !== document; el = el.parentNode) {
-      if (el.tagName === 'PRE') return true
-      if (el.tagName === 'ARTICLE') return false
-    }
-    return false
-  }
+function ABBR({ children }) {
+  return <Tooltip text={children[0]} />
+}
 
-  onTouchStart = e => {
-    this.isCodeBlock = this.isInsideCodeBlock(e.target)
-    this.touchstartX = event.changedTouches[0].screenX
-  }
+ABBR.propTypes = {
+  children: PropTypes.node
+}
 
-  onTouchEnd = () => {
-    this.touchendX = event.changedTouches[0].screenX
-    this.handleSwipeGesture()
-  }
+const renderAst = new rehypeReact({
+  createElement: React.createElement,
+  components: { details: Details, abbr: ABBR }
+}).Compiler
 
-  handleSwipeGesture = () => {
-    if (this.isCodeBlock) return
-    const { prev, next } = this.props
+export default function Markdown({
+  htmlAst,
+  prev,
+  next,
+  tutorials,
+  githubLink
+}) {
+  const touchstartXRef = useRef(0)
+  const touchendXRef = useRef(0)
+  const isCodeBlockRef = useRef(false)
 
-    if (this.touchstartX - this.touchendX > 100) {
+  const handleSwipeGesture = useCallback(() => {
+    if (isCodeBlockRef.current) return
+
+    if (touchstartXRef.current - touchendXRef.current > 100) {
       navigate(next)
     }
 
-    if (this.touchendX - this.touchstartX > 100) {
+    if (touchendXRef.current - touchstartXRef.current > 100) {
       navigate(prev)
     }
-  }
+  }, [prev, next])
 
-  render() {
-    const { markdown, githubLink, prev, next, tutorials } = this.props
+  const onTouchStart = useCallback(e => {
+    isCodeBlockRef.current = isInsideCodeBlock(e.target)
+    touchstartXRef.current = event.changedTouches[0].screenX
+  }, [])
 
-    return (
-      <Content id="markdown-root">
-        {tutorials && (
-          <TutorialsWrapper>
-            <Tutorials tutorials={tutorials} compact={true} />
-          </TutorialsWrapper>
+  const onTouchEnd = useCallback(() => {
+    touchendXRef.current = event.changedTouches[0].screenX
+    handleSwipeGesture()
+  }, [])
+
+  useEffect(() => {
+    document.addEventListener('touchstart', onTouchStart, false)
+    document.addEventListener('touchend', onTouchEnd, false)
+
+    return () => {
+      document.removeEventListener('touchstart', onTouchStart)
+      document.removeEventListener('touchend', onTouchEnd)
+    }
+  }, [])
+
+  return (
+    <Content id="markdown-root">
+      {tutorials && (
+        <TutorialsWrapper>
+          <Tutorials tutorials={tutorials} compact={true} />
+        </TutorialsWrapper>
+      )}
+      <GithubLink href={githubLink} target="_blank" rel="noreferrer noopener">
+        <i /> Edit on GitHub
+      </GithubLink>
+      <div className="markdown-body">{renderAst(htmlAst)}</div>
+      <NavigationButtons>
+        {prev ? (
+          <LocalLink href={prev} as={Button}>
+            <i className="prev" />
+            <span>Prev</span>
+          </LocalLink>
+        ) : (
+          <Button disabled={true}>
+            <i className="prev" />
+            <span>Prev</span>
+          </Button>
         )}
-        <GithubLink href={githubLink} target="_blank" rel="noreferrer noopener">
-          <i /> Edit on GitHub
-        </GithubLink>
-        <ReactMarkdown
-          key={githubLink}
-          className="markdown-body"
-          escapeHtml={false}
-          source={markdown}
-          renderers={{
-            virtualHtml: HtmlRenderer
-          }}
-        />
-        <NavigationButtons>
-          {prev ? (
-            <LocalLink href={prev} as={Button}>
-              <i className="prev" />
-              <span>Prev</span>
-            </LocalLink>
-          ) : (
-            <Button disabled={true}>
-              <i className="prev" />
-              <span>Prev</span>
-            </Button>
-          )}
-          {next ? (
-            <LocalLink href={next} as={Button} disabled={!next}>
-              <span>Next</span>
-              <i className="next" />
-            </LocalLink>
-          ) : (
-            <Button disabled={true}>
-              <span>Next</span>
-              <i className="next" />
-            </Button>
-          )}
-        </NavigationButtons>
-      </Content>
-    )
-  }
+        {next ? (
+          <LocalLink href={next} as={Button} disabled={!next}>
+            <span>Next</span>
+            <i className="next" />
+          </LocalLink>
+        ) : (
+          <Button disabled={true}>
+            <span>Next</span>
+            <i className="next" />
+          </Button>
+        )}
+      </NavigationButtons>
+    </Content>
+  )
 }
 
 Markdown.propTypes = {
-  markdown: PropTypes.string.isRequired,
+  htmlAst: PropTypes.object.isRequired,
   githubLink: PropTypes.string.isRequired,
   tutorials: PropTypes.object,
   prev: PropTypes.string,
